+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 All notable changes to the "confirma-testadapter" extension will be documented in this file.
 
 
-## [0.1.0-beta ?]
+## [Unreleased]
 
 - Initial release
+- compatible with godot-confirma v0.6.1-beta
 
 ## [0.0.0 09.07.2024]
 - begining of the project